refactor: renderProductSelectOptions 함수 분리
