+++ src/basic/main.basic.js
@@ -1,4 +1,4 @@
-import { getProductOptionStyle, getSalesInfoText, getTotalStock, isOutOfStock } from './features/product/service';
+import { renderProductSelectOptions } from './features/product/render';
 
 let itemCount;
 let lastSelector;
@@ -232,7 +232,7 @@ function main() {
   root.appendChild(manualToggle);
   root.appendChild(manualOverlay);
 
-  renderProductSelectOptions();
+  renderProductSelectOptions(selector, productList);
   handleCalculateCartStuff();
 
   setTimeout(() => {
@@ -243,7 +243,7 @@ function main() {
         luckyItem.value = Math.round((luckyItem.originalValue * 80) / 100);
         luckyItem.onSale = true;
         alert('⚡번개세일! ' + luckyItem.name + '이(가) 20% 할인 중입니다!');
-        renderProductSelectOptions();
+        renderProductSelectOptions(selector, productList);
         doUpdatePricesInCart();
       }
     }, 30000);
@@ -267,33 +267,12 @@ function main() {
           alert('💝 ' + suggest.name + '은(는) 어떠세요? 지금 구매하시면 5% 추가 할인!');
           suggest.value = Math.round((suggest.value * (100 - 5)) / 100);
           suggest.suggestSale = true;
-          renderProductSelectOptions();
+          renderProductSelectOptions(selector, productList);
           doUpdatePricesInCart();
         }
       }
     }, 60000);
   }, Math.random() * 20000);
-}
-
-function renderProductSelectOptions() {
-  // 함수 재호출 시 셀렉 옵션 초기화
-  // 초기화하지 않을 시 옵션이 계속해서 추가됨
-  selector.innerHTML = '';
-
-  const totalStock = getTotalStock(productList);
-  selector.style.borderColor = totalStock < 50 ? 'orange' : '';
-
-  const options = productList.map((product) => {
-    const option = document.createElement('option');
-    option.value = product.id;
-    option.textContent = getSalesInfoText(product);
-    option.className = getProductOptionStyle(product);
-    option.disabled = isOutOfStock(product);
-
-    return option;
-  });
-
-  selector.append(...options);
 }
 
 // TODO: doRenderBonusPoints 리팩토링 후 개선 예정
@@ -755,6 +734,6 @@ cartDisp.addEventListener('click', function (event) {
     }
 
     handleCalculateCartStuff();
-    renderProductSelectOptions();
+    renderProductSelectOptions(selector, productList);
   }
 });
